Extraction/Output Operator

Extract a message to output

lhs %>>% rhs

Arguments

lhs
Either a message type identifier like "@info" or an entire message.

rhs
Optional if operator is used with function syntax

Details

This operator provides a programmatic use for logdec outside the context of comments. However, it should be noted that this operator is essentially a wrapper around the engine-specific function call, i.e. cat(), cli::cli_alert_*(), etc.

See also

Examples

# Conventional Use:
"@info" %>>% "Output some form of informational message"
"" %>>% "This will also work, but is a bit odd"

# Functional Use:
`%>>%`("@success Successfully outputted this!")
`%>>%`("@success", "Outputted again!")

# Additional Use:
options(logdec.engine = "cli")
"@info" %>>% "You can change the logdec engine with {.fn {options()}}"

#> ℹ You can change the logdec engine with `TRUE()`, `cairo()`, `xdg-open()`, `FALSE()`, `function (msg) 
#> {
#>     custom_handler <- getOption("cli.default_handler")
#>     if (is.function(custom_handler)) {
#>         custom_handler(msg)
#>     }
#>     else {
#>         cli_server_default(msg)
#>     }
#> }()`, `FALSE()`, `FALSE()`, `1()`, `+ ()`, `c(unordered = "contr.treatment", ordered = "contr.poly")()`, `<pointer: 0x55f55c829fa8>()`, `TRUE()`, `256()`, `TRUE()`, `c("datasets", "utils", "grDevices", "graphics", "stats", "methods")()`, `default()`, `60()`, `function (...) 
#> ragg::agg_png(..., bg = bg)()`, `FALSE()`, `7()`, `character(0)()`, `../articles/()`, `logdec()`, `logdec()`, `%>>%()`, `list(`%>>%` = "%>>%", get_engine_fn = "get_engine_fn", list_engines = "list_engines", output = "output", parse_tag = "parse_tag")()`, `../reference/()`, `dvips()`, `FALSE()`, `vi()`, `native.enc()`, `default()`, `5000()`, `PRE.fansi SPAN {padding-top: .25em; padding-bottom: .25em};()`, `all()`, `8()`, `FALSE()`, `c("bright", "256")()`, `TRUE()`, `c("vignette", "demo", "help")()`, `FALSE()`, `R/4.1.1 R (4.1.1 x86_64-pc-linux-gnu x86_64 linux-gnu) on GitHub Actions()`, `NA()`, `FALSE()`, `2()`, `TRUE()`, `FALSE()`, `TRUE()`, `FALSE()`, `TRUE()`, `cli()`, `mailto()`, `default()`, `99999()`, `TRUE()`, `na.omit()`, `1()`, `50()`, `.()`, `/opt/R/4.1.1/lib/R/bin/pager()`, `a4()`, `NA()`, `FALSE()`, `TRUE()`, `/usr/bin/xdg-open()`, `source()`, `/usr/bin/lpr()`, `> ()`, `c(RSPM = "https://packagemanager.rstudio.com/all/__linux__/focal/latest", CRAN = "https://cloud.r-project.org")()`, ` 	
#> "\'`><=%;,|&{()}()`, `0()`, `TRUE()`, `TRUE()`, `TRUE()`, `TRUE()`, `list(strict.width = "no", digits.d = 3, vec.len = 4, list.len = 99, deparse.lines = NULL, drop.deparse.attr = TRUE, formatNum = function (x, ...) 
#> format(x, trim = TRUE, drop0trailing = TRUE, ...))()`, ``()`, `FALSE()`, `/usr/bin/texi2dvi()`, `60()`, `3()`, `1e-05()`, `FALSE()`, `/usr/bin/unzip()`, `TRUE()`, `FALSE()`, `0()`, `1000()`, and `80()`

"@info" %>>% "As seen above, you can also use {.pkg glue} syntax too!"

#> ℹ As seen above, you can also use glue syntax too!

some_value <- rnorm(1)
"@success" %>>% "Since you can use glue, you can pass variables {some_value}."

#> Error in eval(expr, envir = list(`?` = function(...) stop()), enclos = envir): object 'some_value' not found